+++ app/code/community/Nosto/Tagging/Block/Variation.php
@@ -64,7 +64,7 @@ class Nosto_Tagging_Block_Variation extends Mage_Core_Block_Template
         $dataHelper = Mage::helper('nosto_tagging');
         if ($dataHelper->isMultiCurrencyMethodExchangeRate(Mage::app()->getStore())) {
             return Mage::app()->getStore()->getCurrentCurrencyCode();
-        } else if ($dataHelper->isVariationEnabled(Mage::app()->getStore())){
+        } elseif ($dataHelper->isVariationEnabled(Mage::app()->getStore())) {
             $groupId = Mage::getSingleton('customer/session')->getCustomerGroupId();
 
             /** @var Mage_Customer_Model_Group $customerGroup */
@@ -82,13 +82,18 @@ class Nosto_Tagging_Block_Variation extends Mage_Core_Block_Template
     /**
      * Tells if store uses multiple currencies
      *
-     * @return string
+     * @return bool
      */
-    public function useMultiCurrency()
+    public function useMultiCurrencyOrPriceVariation()
     {
         /** @var Nosto_Tagging_Helper_Data $helper */
         $helper = Mage::helper('nosto_tagging');
 
-        return $helper->isMultiCurrencyMethodExchangeRate(Mage::app()->getStore());
+        $enabled = $helper->isMultiCurrencyMethodExchangeRate(Mage::app()->getStore());
+        if (!$enabled) {
+            $enabled = $helper->isVariationEnabled(Mage::app()->getStore());
+        }
+
+        return $enabled;
     }
 }