+++ app/code/community/Nosto/Tagging/Helper/Price.php
@@ -101,13 +101,7 @@ class Nosto_Tagging_Helper_Price extends Mage_Core_Helper_Abstract
 
         switch ($product->getTypeId()) {
             case Mage_Catalog_Model_Product_Type::TYPE_BUNDLE:
-                // Get the bundle product "from" / min price.
-                // Price for bundled "parent" product cannot be configured in
-                // store admin. In practise there is no such thing as
-                // parent product for the bundled type product
-                /** @var Mage_Bundle_Model_Product_Price $model */
-                $model = $product->getPriceModel();
-                $price = $model->getTotalPrices($product, 'min', $inclTax);
+                $price = $this->getBundleProductPrices($product, $finalPrice, $inclTax);
                 break;
             case Mage_Catalog_Model_Product_Type::TYPE_GROUPED:
                 // Get the grouped product "starting at" price.
@@ -158,9 +152,11 @@ class Nosto_Tagging_Helper_Price extends Mage_Core_Helper_Abstract
                         $productModel = Mage::getModel('catalog/product')->load(
                             $associatedProduct->getId()
                         );
-                        $variationPrice = $this->_getProductPrice($productModel, $finalPrice, $inclTax);
-                        if (!$lowestPrice || $variationPrice < $lowestPrice) {
-                            $lowestPrice = $variationPrice;
+                        if ($productModel && $productModel->isAvailable()) {
+                            $variationPrice = $this->_getProductPrice($productModel, $finalPrice, $inclTax);
+                            if (!$lowestPrice || $variationPrice < $lowestPrice) {
+                                $lowestPrice = $variationPrice;
+                            }
                         }
                     }
                     $price = $lowestPrice;
@@ -176,6 +172,74 @@ class Nosto_Tagging_Helper_Price extends Mage_Core_Helper_Abstract
         }
 
         return $price;
+    }
+
+    /**
+     * @param Mage_Catalog_Model_Product $product
+     * @param bool $finalPrice
+     * @param bool $inclTax
+     * @return float
+     * @suppress PhanUndeclaredMethod
+     */
+    public function getBundleProductPrices(
+        Mage_Catalog_Model_Product $product,
+        $finalPrice = false,
+        $inclTax = true
+    )
+    {
+        // Get the bundle product "from" / min price.
+        // Price for bundled "parent" product cannot be configured in
+        // store admin. In practise there is no such thing as
+        // parent product for the bundled type product
+        /** @var Mage_Bundle_Model_Product_Price $model */
+        $model = $product->getPriceModel();
+        $minBundlePrice = $model->getTotalPrices($product, 'min', $inclTax, $finalPrice);
+
+        if ($finalPrice) {
+
+            return $minBundlePrice;
+        }
+
+        /** @var Mage_Bundle_Model_Product_Type $typeInstance */
+        $typeInstance = $product->getTypeInstance();
+        $typeInstance->setStoreFilter($product->getStoreId(), $product);
+
+        /** @var Mage_Bundle_Model_Resource_Option_Collection $optionCollection */
+        $optionCollection = $typeInstance->getOptionsCollection($product);
+
+        $selectionCollection = $typeInstance->getSelectionsCollection(
+            $typeInstance->getOptionsIds($product),
+            $product
+        );
+
+        $options = $optionCollection->appendSelections(
+            $selectionCollection,
+            false,
+            Mage::helper('catalog/product')->getSkipSaleableCheck()
+        );
+        $sumListPrice = 0;
+        /** @var Mage_Bundle_Model_Option $option */
+        foreach ($options as $option){
+            $selections  = $option->getSelections();
+            $minSimpleProductPricePrice = null;
+            $simpleProductListPrice = null;
+            /**
+             * @var Mage_Catalog_Model_Product $selection
+             */
+            foreach ($selections as $selection){
+                if ($selection->isAvailable()) {
+                    $simpleProductPrice = $this->_getProductPrice($selection, true, $inclTax);
+                    if ($minSimpleProductPricePrice === null || $simpleProductPrice < $minSimpleProductPricePrice) {
+                        $minSimpleProductPricePrice = $simpleProductPrice;
+                        $simpleProductListPrice = $this->_getProductPrice($selection, false, $inclTax);
+                    }
+                }
+            }
+
+            $sumListPrice += $simpleProductListPrice;
+        }
+
+        return max($sumListPrice, $minBundlePrice);
     }
 
     /**
@@ -296,7 +360,11 @@ class Nosto_Tagging_Helper_Price extends Mage_Core_Helper_Abstract
      * @param $isFinalPrice true means it is final price, or it is list price
      * @return float the price
      */
-    public function getProductTaggingPrice(Mage_Catalog_Model_Product $product, Mage_Core_Model_Store $store, $isFinalPrice)
+    public function getProductTaggingPrice(
+        Mage_Catalog_Model_Product $product,
+        Mage_Core_Model_Store $store,
+        $isFinalPrice
+    )
     {
         $basePrice = $this->getDisplayPriceInStore($product, $store, $isFinalPrice);