+++ app/code/community/Nosto/Tagging/Helper/Url.php
@@ -36,6 +36,9 @@
  */
 class Nosto_Tagging_Helper_Url extends Mage_Core_Helper_Abstract
 {
+    const URL_PATH_NOSTO_CONFIG = 'adminhtml/system_config/edit/section/nosto_tagging';
+    const MAGENTO_URL_OPTION_STORE_CODE = 'store';
+
     /**
      * The ___store parameter in Magento URLs
      */
@@ -283,8 +286,8 @@ class Nosto_Tagging_Helper_Url extends Mage_Core_Helper_Abstract
             self::MAGENTO_PATH_CART,
             $defaultParams
         );
-        if (count($additionalParams) > 0) {
-            foreach ($additionalParams as $key=>$val) {
+        if (!empty($additionalParams)) {
+            foreach ($additionalParams as $key => $val) {
                 $url = Nosto_Request_Http_HttpRequest::replaceQueryParamInUrl(
                     $key,
                     $val,
@@ -409,25 +412,10 @@ class Nosto_Tagging_Helper_Url extends Mage_Core_Helper_Abstract
      */
     public function removeQueryParamFromUrl($url, $param)
     {
-        $modifiedUrl = $url;
-        $urlParts = Nosto_Request_Http_HttpRequest::parseUrl($url);
-        if (
-            is_array($urlParts)
-            && isset($urlParts['query'])
-        ) {
-            $queryArray = Nosto_Request_Http_HttpRequest::parseQueryString($urlParts['query']);
-            if (isset($queryArray[$param])) {
-                unset($queryArray[$param]);
-                if (empty($queryArray)) {
-                    unset($urlParts['query']);
-                } else {
-                    $urlParts['query'] = http_build_query($queryArray);
-                }
-                $modifiedUrl = Nosto_Request_Http_HttpRequest::buildUrl($urlParts);
-            }
-        }
+        $zendUrl = Zend_Uri_Http::fromString($url);
+        $zendUrl->removeQueryParameters([$param]);
 
-        return $modifiedUrl;
+        return $zendUrl->getUri();
     }
 
     /**
@@ -540,5 +528,23 @@ class Nosto_Tagging_Helper_Url extends Mage_Core_Helper_Abstract
         );
 
         return $url;
+    }
+
+    /**
+     * Gets the absolute URL to the Nosto configuration page
+     *
+     * @param Mage_Core_Model_Store $store the store to get the url for.
+     *
+     * @return string the url.
+     */
+    public function getAdminNostoConfiguratioUrl(Mage_Core_Model_Store $store)
+    {
+        $params = array(
+            self::MAGENTO_URL_OPTION_STORE_CODE => $store->getCode()
+        );
+        /** @var Mage_Adminhtml_Helper_Data $adminHtmlHelper */
+        $adminHtmlHelper = Mage::helper('adminhtml');
+
+        return $adminHtmlHelper->getUrl(self::URL_PATH_NOSTO_CONFIG, $params);
     }
 }