+++ app/code/community/Nosto/Tagging/controllers/AddToCartController.php
@@ -93,7 +93,10 @@ class Nosto_Tagging_AddToCartController extends Mage_Checkout_CartController
 
             if (!$this->_getSession()->getNoCartRedirect(true)) {
                 if (!$cart->getQuote()->getHasError()) {
-                    $message = $this->__('%s was added to your shopping cart.', Mage::helper('core')->escapeHtml($product->getName()));
+                    $message = $this->__(
+                        '%s was added to your shopping cart.',
+                        Mage::helper('core')->escapeHtml($product->getName())
+                    );
                     $this->_getSession()->addSuccess($message);
                 }
                 $this->_goBack();